Job launch step
