+++ src/tapis-ui/_wrappers/Wizard/Wizard.module.scss
@@ -25,6 +25,10 @@
   }
 }
 
+.step {
+  padding-top: 1em;
+}
+
 .controls {
   display: flex;
   justify-content: flex-end;
