+++ src/tapis-ui/components/jobs/JobLauncher/JobLauncherWizard.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { WizardStep } from 'tapis-ui/_wrappers/Wizard';
 import { Wizard } from 'tapis-ui/_wrappers';
 import * as Jobs from '@tapis/tapis-typescript-jobs';
-import { JobBasics, JobBasicsSummary } from './steps/JobBasics';
+import { JobStart, JobStartSummary } from './steps/JobStart';
 import { FileInputs, FileInputsSummary } from './steps/FileInputs';
 import { ExecSystem, ExecSystemSummary } from './steps/ExecSystem';
 import * as Apps from '@tapis/tapis-typescript-apps';
@@ -14,22 +14,22 @@ type JobLauncherWizardProps = {
 const JobLauncherWizard: React.FC<JobLauncherWizardProps> = ({ app }) => {
   const steps: Array<WizardStep> = [
     {
-      id: 'required',
-      name: 'Job Stuff',
-      render: <JobBasics app={app} />,
-      summary: <JobBasicsSummary />,
-    },
-    {
-      id: 'fileInputs',
-      name: 'File Stuff',
-      render: <FileInputs app={app} />,
-      summary: <FileInputsSummary />,
+      id: 'start',
+      name: `App Selection`,
+      render: <JobStart app={app} />,
+      summary: <JobStartSummary />,
     },
     {
       id: 'execSystem',
       name: 'Execution System',
       render: <ExecSystem app={app} />,
       summary: <ExecSystemSummary />
+    },
+    {
+      id: 'fileInputs',
+      name: 'File Stuff',
+      render: <FileInputs app={app} />,
+      summary: <FileInputsSummary />,
     }
   ];
 
@@ -37,6 +37,7 @@ const JobLauncherWizard: React.FC<JobLauncherWizardProps> = ({ app }) => {
     name: `${app.id}-${app.version}-${new Date().toISOString().slice(0, -5)}`,
     appId: app.id,
     appVersion: app.version,
+    execSystemId: app.jobAttributes?.execSystemId
   };
 
   return (
